Added support for UTF-8 (for Windows)
The el... ekhm... the scrolls I've found in the dungeons were very hard to decode. Lots of unknown characters, strange signs and bad writing style... After >days< of laborious efforts and nights spent in libraries, with a help of local wizards, I finally could understand the message. The scrolls need to be rewritten now.